// src/app/generated_services/model/carlonGracieBackendFinancesApplicationDTOsShowContractDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarlonGracieBackendSharedDomainEnumsContractStatus } from './carlonGracieBackendSharedDomainEnumsContractStatus';


export interface CarlonGracieBackendFinancesApplicationDTOsShowContractDTO { 
    id?: string;
    studentId?: string;
    feePlanId?: string;
    feePlanName?: string | null;
    feePlanMonthDuration?: number;
    startDate?: string;
    endDate?: string;
    monthlyAmount?: number;
    status?: CarlonGracieBackendSharedDomainEnumsContractStatus;
    notes?: string | null;
    createdAt?: string;
    totalFees?: number;
    paidFees?: number;
    pendingFees?: number;
    overdueFees?: number;
}
export namespace CarlonGracieBackendFinancesApplicationDTOsShowContractDTO {
}