// src/app/generated_services/model/carlonGracieBackendSharedDomainEnumsContractStatus.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CarlonGracieBackendSharedDomainEnumsContractStatus = 'Active' | 'Inactive' | 'Suspended' | 'Terminated' | 'Cancelled' | 'Expired';

export const CarlonGracieBackendSharedDomainEnumsContractStatus = {
    Active: 'Active' as CarlonGracieBackendSharedDomainEnumsContractStatus,
    Inactive: 'Inactive' as CarlonGracieBackendSharedDomainEnumsContractStatus,
    Suspended: 'Suspended' as CarlonGracieBackendSharedDomainEnumsContractStatus,
    Terminated: 'Terminated' as CarlonGracieBackendSharedDomainEnumsContractStatus,
    Cancelled: 'Cancelled' as CarlonGracieBackendSharedDomainEnumsContractStatus,
    Expired: 'Expired' as CarlonGracieBackendSharedDomainEnumsContractStatus
};